playwright/tests: Move Compliance lint warning test to right suite
The test was placed into boot tests even though it is not a boot test.
Moved to Customizations test suite.
